+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/block_entity/DnaExtractorBlockEntity.kt
@@ -31,7 +31,7 @@ class DnaExtractorBlockEntity(
 			return when (slot) {
 				INPUT_SLOT_INDEX -> stack.`is`(ModItems.CELL) || stack.`is`(ModItems.GMO_CELL)
 				OVERCLOCK_SLOT_INDEX -> stack.`is`(ModItems.OVERCLOCKER)
-				OUTPUT_SLOT_INDEX -> false
+				OUTPUT_SLOT_INDEX -> true
 				else -> false
 			}
 		}